fix IdnaTest.inc generation w/out of tree builds
The current rule will always look for IdnaTest.txt in the current
directory (and output IdnaTest.inc to the current directory).
When doing an out-of-tree build, the Makefile (via vpaths) will
find the source files, but when the rule runs, it won't.  Change
the rules to use automatic makefile variables so it can deal with
srcdir/builddir dynamically.

It does mean that always write IdnaTest.txt to the builddir and
not the sourcedir, but I think that's what the code intends.
When doing a `make dist`, it'll source the right file and save
it to the right place in the tarball.