apply TR#46 non-transitional processing by default
This switches the default library behavior from "pure" IDNA2008
to IDNA2008 + non-transitional processing. This can be API and
ABI break for certain applications which intentionally needed
the non amended IDNA2008.  It is done in the interest of interoperability
based on the fact that this is what most application writers care about
rather than strict compliance with a particular protocol.

Resolves #49

Signed-off-by: Nikos Mavrogiannopoulos <nmav@gnutls.org>